refactor(core): group core-side browser modules under core/browser
The five `browser*`-prefixed modules in `src/core/` were only grouped by
their filename prefix. Move them into `src/core/browser/` and drop the
now-redundant prefix:

  browserLoader.ts        -> browser/loader.ts
  browserOnlyRun.ts       -> browser/onlyRun.ts
  browserRunPlanner.ts    -> browser/runPlanner.ts
  browserWatchControls.ts -> browser/watchControls.ts
  browserGlobalSetup.ts   -> browser/globalSetupStage.ts

The globalSetup module keeps a distinguishing suffix: a bare
`browser/globalSetup.ts` would shadow the node `core/globalSetup.ts` at a
glance, and `onlyRun.ts` imports from both.

`isBrowserProject.ts` deliberately stays put — it is the shared routing
predicate the node-path modules (plan resolution, environment grouping,
Rsbuild setup, node executor) read, not browser-mode implementation.

Pure file move plus import-path updates; no behavior change.
