+++ packages/core/src/core/browser/loader.ts
@@ -6,15 +6,15 @@ import type {
   ProjectContext,
   RstestContext,
   TestExecutor,
-} from '../types';
-import type { CoverageProvider } from '../types/coverage';
-import { color, logger } from '../utils';
+} from '../../types';
+import type { CoverageProvider } from '../../types/coverage';
+import { color, logger } from '../../utils';
 
 export type {
   BrowserTestRunOptions,
   BrowserTestRunResult,
   ListBrowserTestsOptions,
-} from '../types';
+} from '../../types';
 
 /**
  * The subset of {@link BrowserTestRunOptions} that configures a browser
